Big FDB Update, Minor DB Change
Added Several new functions to FDB.Java that populate UI fields, added search methods to search with a variety of arguments. updated the database so that players with a single name have a "-" in place for their last name so that it doesnt look errornous in the UI table, and also as it was bugging out the search by last name.